Refactor remote bitbang code.
